Automatically select appropriate model for IMU

Runs through a model selection algorithm to determine the best model

auto.imu(data, model = 3 * AR1() + WN() + RW() + QN() + DR(), bootstrap = F,
  alpha = 0.05, robust = F, eff = 0.6, B = 50, G = 1e+06,
  seed = 1337)

Arguments

data

A vector, matrix, data.frame, or imu object with either 1, 3, or 6 columns.

model

A ts.model object that is the largest model to be tested.

bootstrap

A bool that is either true or false to indicate whether we use bootstrap or asymptotic By default, we use asymptotic.

alpha

A double that indicates the level of confidence for the WV CI.

robust

A boolean that indicates whether to use robust estimation.

eff

A double between 0 and 1 that indicates the efficiency for the robust estimation.

B

A integer that contains the amount of bootstrap replications

G

A integer that indicates the amount of guesses for caliberating the startup.

seed

A integer that controls the reproducibility of the auto model selection phase.

Value

A auto.imu object.

Details

The auto.imu object stores two important features for each signal:

[[1]]A matrix containing model output

[[2]]The best gmwm object.

To access it for each signal use: object[[i]][[1]] or object[[i]][[2]], where \(i\) denotes the signal.

Examples

## Not run: ------------------------------------
# if(!require("imudata")){
# install_imudata()
# library("imudata")
# }
# 
# data(imu6)
# 
# # Example 1
# test1 = imu(imu6, gyros = 1:3, accels = NULL, axis = c('X', 'Y', 'Z'), freq = 100)
# 
# m = auto.imu(test1)
# 
# # Process 1's model table
# m[[1]][[1]]
# 
# # Process 1's best fitting gmwm object
# m[[1]][[2]]
# 
## ---------------------------------------------

Author
JJB