GMWM for (Robust) Sensor

GMM object

gmwm.imu(model, data, compute.v = "fast", robust = F, eff = 0.6, ...)

Arguments

model

A ts.model object containing one of the allowed models.

data

A matrix or data.frame object with only column (e.g. \(N \times 1\)), or a lts object, or a gts object.

compute.v

A string indicating the type of covariance matrix solver. "fast", "bootstrap", "asymp.diag", "asymp.comp", "fft"

robust

A boolean indicating whether to use the robust computation (TRUE) or not (FALSE).

eff

A double between 0 and 1 that indicates the efficiency.

...

Other arguments passed to the main gmwm function

Value

A gmwm object with the structure:

estimate

Estimated Parameters Values from the GMWM Procedure

init.guess

Initial Starting Values given to the Optimization Algorithm

wv.empir

The data's empirical wavelet variance

ci.low

Lower Confidence Interval

ci.high

Upper Confidence Interval

orgV

Original V matrix

V

Updated V matrix (if bootstrapped)

omega

The V matrix inversed

obj.fun

Value of the objective function at Estimated Parameter Values

theo

Summed Theoretical Wavelet Variance

decomp.theo

Decomposed Theoretical Wavelet Variance by Process

scales

Scales of the GMWM Object

robust

Indicates if parameter estimation was done under robust or classical

eff

Level of efficiency of robust estimation

model.type

Models being guessed

compute.v

Type of V matrix computation

augmented

Indicates moments have been augmented

alpha

Alpha level used to generate confidence intervals

expect.diff

Mean of the First Difference of the Signal

N

Length of the Signal

G

Number of Guesses Performed

H

Number of Bootstrap replications

K

Number of V matrix bootstraps

model

ts.model supplied to gmwm

model.hat

A new value of ts.model object supplied to gmwm

starting

Indicates whether the procedure used the initial guessing approach

seed

Randomization seed used to generate the guessing values

freq

Frequency of data

Details

This version of the gmwm function has customized settings ideal for modeling with an IMU object. If you seek to model with an Gauss Markov, GM, object. Please note results depend on the freq specified in the data construction step within the imu. If you wish for results to be stable but lose the ability to interpret with respect to freq, then use AR1 terms.

Examples

## Not run: ------------------------------------
# # Example data generation
# data = gen.gts(GM(beta=0.25,sigma2_gm=1),10000, freq = 5)
# results = gmwm.imu(GM(),data)
# inference = summary(results)
# 
# # Example with IMU Data
# 
# 
# 
## ---------------------------------------------