Add LaTeX markup to R data.

Add LaTeX markup to R data.

latex_markup(df, ...)

Arguments

df
The dataframe, data.table, tibble, etc, containing the data to add LaTeX markup to.

...
One or more unquoted, comma separated, variable names, with quoted Latex environment names (no backslashes or curly braces)/ #@param markup_style

Examples

not_run({

  x <- latex_markup(iris, Species="section", Petal.Width="subsection*")#, markup_style="section")

   # >  head(x)
   #    Sepal.Length Sepal.Width Petal.Length       Petal.Width           Species
   #  1          5.1         3.5          1.4 \\subsection*{0.2} \\section{setosa}
   #  2          4.9         3.0          1.4 \\subsection*{0.2} \\section{setosa}
   #  3          4.7         3.2          1.3 \\subsection*{0.2} \\section{setosa}
   #  4          4.6         3.1          1.5 \\subsection*{0.2} \\section{setosa}
   #  5          5.0         3.6          1.4 \\subsection*{0.2} \\section{setosa}
   #  6          5.4         3.9          1.7 \\subsection*{0.4} \\section{setosa}

})