Methods for working with mrgsims objects.

These methods help the user view simulation output and extract simulated data to work with further. The methods listed here for the most part have generics defined by R or other R packages. See the seealso section for other methods defined by mrgsolve that have their own documentation pages.

# S4 method for mrgsims
$(x, name)

# S4 method for mrgsims
tail(x, ...)

# S4 method for mrgsims
head(x, ...)

# S4 method for mrgsims
dim(x)

# S4 method for mrgsims
names(x)

# S4 method for mrgsims
as.data.frame(x, row.names = NULL, optional = FALSE,
  ...)

# S4 method for mrgsims
as.matrix(x, ...)

# S4 method for mrgsims
subset(x, ...)

# S4 method for mrgsims
summary(object, ...)

# S4 method for mrgsims
show(object)

Arguments

x
mrgsims object

name
name of column of simulated output to retain

...
passed to other functions

row.names
passed to as.data.frame

optional
passed to as.data.frame

object
passed to show

Details

Most methods should behave as expected according to other method commonly used in R (e.g. head, tail, as.data.frame, etc ...)

subset coreces simulated output to data.frame and passes to subset.data.frame

$ selects a column in the simulated data and returns numeric

head see head.matrix; returns simulated data

tail see tail.matrix; returns simulated data

dim, nrow, ncol returns dimensions, number of rows, and number of columns in simulated data

as.data.frame coreces simulated data to data.frame and returns the data.frame

as.matrix returns matrix of simulated data

as.tbl coreces simulated to tbl_df; requires dplyr

summary coreces simulated data to data.frame and passes to summary.data.frame

plot plots simulated data; see plot_mrgsims

See also

Examples

## example("mrgsims")

mod <- mrgsolve:::house() %>% init(GUT=100)

#> Warning: cannot open compressed file '/Users/kyleb/git/m4solve/pkg/project/housemodel.RDS', probable reason 'No such file or directory'

#> Error in gzfile(file, "rb"): cannot open the connection

out <- mrgsim(mod)

#> Error in mrgsim(mod): no slot of name "args" for this object of class "standardGeneric"

class(out)

#> Error in eval(expr, envir, enclos): object 'out' not found

out

#> Error in eval(expr, envir, enclos): object 'out' not found

head(out)

#> Error in head(out): object 'out' not found

tail(out)

#> Error in tail(out): object 'out' not found

mrgsolve:::mod(out)

#> Error in mrgsolve:::mod(out): object 'out' not found

dim(out)

#> Error in eval(expr, envir, enclos): object 'out' not found

names(out)

#> Error in eval(expr, envir, enclos): object 'out' not found

mat <- as.matrix(out)

#> Error in as.matrix(out): object 'out' not found

df <- as.data.frame(out)

#> Error in as.data.frame(out): object 'out' not found

df <- subset(out, time < 12) ## a data frame

#> Error in subset(out, time < 12): object 'out' not found

out$CP

#> Error in eval(expr, envir, enclos): object 'out' not found

plot(out)

#> Error in plot(out): object 'out' not found

plot(out, CP~.)

#> Error in plot(out, CP ~ .): object 'out' not found

plot(out, CP+RESP~time, scales="same", xlab="Time", main="Model sims")

#> Error in plot(out, CP + RESP ~ time, scales = "same", xlab = "Time", main = "Model sims"): object 'out' not found